feat: persist ml 3-value awareness via connaissance_ml enum
